+++ docs/requirements.txt
@@ -7,7 +7,8 @@ black==19.10b0; python_version >= "3.6"
 certifi==2020.12.5; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
 chardet==4.0.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
 click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
-colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6") or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6")
+cloudpickle==2.0.0; python_version >= "3.6" and python_version < "3.9"
+colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6") and python_version < "3.9" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6") and python_version < "3.9" and platform_system == "Windows"
 coverage==5.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
 cycler==0.10.0; python_version >= "3.6"
 cython==3.0a6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
@@ -19,12 +20,14 @@ importlib-metadata==3.10.1; python_version < "3.8" and python_version >= "3.6" a
 iniconfig==1.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
 isort==4.3.21; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
 jinja2==2.11.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
-joblib==1.0.1; python_version >= "3.6"
+joblib==1.0.1; python_version >= "3.6" and python_version < "3.9"
 kiwisolver==1.3.1; python_version >= "3.6"
+llvmlite==0.34.0; python_version >= "3.6" and python_version < "3.9"
 markupsafe==1.1.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
 matplotlib==3.3.4; python_version >= "3.6"
-numpy==1.19.5; python_version >= "3.6" and python_full_version >= "3.6.1"
-packaging==21.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
+numba==0.51.2; python_version >= "3.6" and python_version < "3.9"
+numpy==1.19.5; python_version >= "3.6" and python_full_version >= "3.6.1" and python_version < "3.9"
+packaging==21.3; python_version >= "3.6" and python_version < "3.9" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6")
 pandas==1.1.5; python_full_version >= "3.6.1"
 pathspec==0.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
 pillow==8.4.0; python_version >= "3.6"
@@ -36,13 +39,15 @@ pytest-cov==2.11.1; (python_version >= "2.7" and python_full_version < "3.0.0") 
 pytest-forked==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
 pytest-xdist==2.3.0; python_version >= "3.6"
 pytest==6.2.3; python_version >= "3.6"
-python-dateutil==2.8.1; python_full_version >= "3.6.1" and python_version >= "3.6"
-pytz==2021.1; python_full_version >= "3.6.1" and python_version >= "3.5"
+python-dateutil==2.8.1; python_full_version >= "3.6.1" and python_version >= "3.6" and python_version < "3.9"
+pytz==2021.1; python_full_version >= "3.6.1" and python_version >= "3.5" and python_version < "3.9"
 regex==2021.4.4; python_version >= "3.6"
 requests==2.25.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
 scikit-learn==0.24.1; python_version >= "3.6"
-scipy==1.5.4; python_version >= "3.6"
+scipy==1.5.4; python_version >= "3.6" and python_version < "3.9"
+shap==0.40.0; python_version < "3.9"
 six==1.15.0; python_full_version >= "3.6.1" and python_version >= "3.6"
+slicer==0.0.7; python_version >= "3.6" and python_version < "3.9"
 snowballstemmer==2.1.0; python_version >= "3.5"
 sphinx-rtd-theme==0.4.3
 sphinx==3.5.4; python_version >= "3.5"
@@ -52,8 +57,9 @@ sphinxcontrib-htmlhelp==1.0.3; python_version >= "3.5"
 sphinxcontrib-jsmath==1.0.1; python_version >= "3.5"
 sphinxcontrib-qthelp==1.0.3; python_version >= "3.5"
 sphinxcontrib-serializinghtml==1.1.4; python_version >= "3.5"
-threadpoolctl==2.1.0; python_version >= "3.6"
+threadpoolctl==2.1.0; python_version >= "3.6" and python_version < "3.9"
 toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
+tqdm==4.62.3; python_version >= "2.7" and python_full_version < "3.0.0" and python_version < "3.9" or python_full_version >= "3.4.0" and python_version < "3.9"
 typed-ast==1.4.3; python_version >= "3.6"
 typing-extensions==3.7.4.3; python_version < "3.8" and python_version >= "3.6"
 urllib3==1.26.6; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.5"
