+++ tests/test_tools.py
@@ -1,43 +1,13 @@
 import os
 import sys
-from contextlib import contextmanager
 
 import pytest
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.ensemble import RandomForestRegressor
-from sklearn.tree._tree import Tree as SKTree
 
 from skranger.ensemble import RangerForestClassifier
 from skranger.ensemble import RangerForestRegressor
-from skranger.tree._tree import Tree
-
-
-# region trick shap into thinking skranger objects are sklearn objects
-# so that shap will process them.
-def getclass(klass):
-    def __class__override(self, item):
-        if item == "__class__":
-            return klass
-        return object.__getattribute__(self, item)
-
-    return __class__override
-
-
-@contextmanager
-def patch_sklearn():
-    tree_orig = Tree.__getattribute__
-    Tree.__getattribute__ = getclass(SKTree)
-    reg_orig = RangerForestRegressor.__getattribute__
-    RangerForestRegressor.__getattribute__ = getclass(RandomForestRegressor)
-    cls_orig = RangerForestClassifier.__getattribute__
-    RangerForestClassifier.__getattribute__ = getclass(RandomForestClassifier)
-    yield
-    Tree.__getattribute__ = tree_orig
-    RangerForestRegressor.__getattribute__ = reg_orig
-    RangerForestClassifier.__getattribute__ = cls_orig
-
-
-# endregion
+from skranger.utils import shap_patch
 
 
 @pytest.mark.skip()
@@ -68,9 +38,9 @@ def test_shap_regressor(boston_X, boston_y):
     forest = RangerForestRegressor(enable_tree_details=True)
     forest.fit(boston_X, boston_y)
 
-    with patch_sklearn():
+    with shap_patch():
         explainer = TreeExplainer(model=forest)
-        shap_values = explainer.shap_values(boston_X)
+    shap_values = explainer.shap_values(boston_X)
     print(shap_values)
 
 
@@ -83,9 +53,9 @@ def test_shap_classifier(iris_X, iris_y):
     forest = RangerForestClassifier(enable_tree_details=True)
     forest.fit(iris_X, iris_y)
 
-    with patch_sklearn():
+    with shap_patch():
         explainer = TreeExplainer(model=forest)
-        shap_values = explainer.shap_values(iris_X)
+    shap_values = explainer.shap_values(iris_X)
     print(shap_values)
